pyproject.toml
[project]
name = "django-porfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.97",
    "django-cors-headers>=4.9.0",
    "django>=6.0.4",
    "djangorestframework>=3.17.1",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
]
